// src/app/models/order.interface.ts
export interface OrderI {
    id?: string;
    cliente?: string;
    clienteId?: string;
    driverID?: string;
    DName?: string;
    super?: string;
    otro?: string;
    listImage?: any;
    cantidad?: number;
    nombre?: string;
    nombreV?: string;
    phone?: string;
    precio?: number;
    ubicacion?: string;
    clienteub?: string;
    estado?: number;
}

    